move ping feature to "feature.ping" package
